new parameter - percentage of autonomus cars on a road

+++ src/main/java/pl/zablocki/core/road/Road.java
@@ -17,7 +17,7 @@ public class Road {
     public int vehicleCounter;
     @Getter
     @Setter
-    private int autonomusPercentage;
+    private double autonomusPercentage;
 
     public Road(int id) {
         this.id = id;
